# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.ResponseModel:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
...
  service.Actor:
    type: object
  service.AddActorsByFilmParams:
    properties:
      actors:
        items:
          type: string
        type: array
      film:
        type: string
    type: object
  service.AddFilmsByActorParams:
    properties:
      actor:
        type: string
      films:
        items:
          type: string
        type: array
    type: object
  service.DeleteActorFilmParams:
    properties:
      actor:
        type: string
      film:
        type: string
    type: object
  service.Film:
    type: object
  service.ResponseModel:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server film library server.
  title: Film Library App API
  version: "1.0"
paths:
  /actor/add:
    post:
      consumes:
      - application/json
      description: Add actor
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: actor data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.Actor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.ResponseModel'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: CreateActor
      tags:
      - actor
  /actor/delete/{actor_name}:
    delete:
      consumes:
      - application/json
      description: Delete actor
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: actor name
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ResponseModel'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: DeleteActor
      tags:
      - actor
  /actor/get/{actor_name}:
    get:
      consumes:
      - application/json
      description: Get actor
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: actor name
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Actor'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: GetActor
      tags:
      - actor
  /actor/get_all:
    get:
      consumes:
      - application/json
      description: Get actors
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sort
        in: header
        name: Sort
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.Actor'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: GetActors
      tags:
      - actor
  /actor/search/{actor_name}:
    get:
      consumes:
      - application/json
      description: Search actor
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: pattern
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: SearchActor
      tags:
      - actor
  /actor/update/{actor_name}:
    patch:
      consumes:
      - application/json
      description: Update actor
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: actor name
        in: query
        name: q
        type: string
      - description: actor data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.Actor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ResponseModel'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: UpdateActor
      tags:
      - actor
...
  /film/add:
    post:
      consumes:
      - application/json
      description: Add film
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: film data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.Film'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ResponseModel'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: CreateFilm
      tags:
      - film
  /film/delete/{film_name}:
    delete:
      consumes:
      - application/json
      description: Delete film
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: film name
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ResponseModel'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: DeleteFilm
      tags:
      - film
  /film/get/{actor_name}:
    get:
      consumes:
      - application/json
      description: Get film
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: film name
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Film'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: GetFilm
      tags:
      - film
  /film/get_all:
    get:
      consumes:
      - application/json
      description: Get films
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sort
        in: header
        name: Sort
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.Film'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: GetFilms
      tags:
      - film
  /film/search/{film_name}:
    get:
      consumes:
      - application/json
      description: Search films
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: pattern
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: SearchFilms
      tags:
      - film
  /film/update/{film_name}:
    patch:
      consumes:
      - application/json
      description: Update film
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: film name
        in: query
        name: q
        type: string
      - description: new film data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.Film'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ResponseModel'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: UpdateFilm
      tags:
      - film
  /relation/actors_by_film:
    post:
      consumes:
      - application/json
      description: Add actors by film
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: relation data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.AddActorsByFilmParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ResponseModel'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: AddActorsByFilm
      tags:
      - relation
  /relation/delete:
    delete:
      consumes:
      - application/json
      description: Delete relation actor film
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: relation data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.DeleteActorFilmParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ResponseModel'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: DeleteActorFilm
      tags:
      - relation
  /relation/films_by_actor:
    post:
      consumes:
      - application/json
      description: Add films by actor
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: relation data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.AddFilmsByActorParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ResponseModel'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: AddFilmsByActor
      tags:
      - relation
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"